+++ R/heeringa.R
@@ -3,7 +3,7 @@
 # mapping dimensions of MDS onto RG
 # ========
 
-heeringa <- function(dist, power = 0.5, mapping = c(1,2,3) ) {
+heeringa <- function(dist, power = 0.5, mapping = c(1,2,3), method = "eigs", center = NULL) {
 
   # check mapping
   if (length(mapping) != 3
@@ -12,10 +12,31 @@ heeringa <- function(dist, power = 0.5, mapping = c(1,2,3) ) {
     stop("mapping is not correct")
   }
 
-  mds <- cmdscale(dist, k = 3)
+  # check format of dist
+  if (class(dist)[1] == "dist") {
+    dist <- as.matrix(dist)
+  }
+
+  if (method == "mds") {
+    mds <- cmdscale(dist, k = 3)
+  } else if (method == "eigs") {
+    e <- RSpectra::eigs(max(dist)-dist, 4)
+    d <- diag(e$values[2:4])^0.5
+    mds <- e$vectors[,2:4] %*% d
+  }
+
+  if (is.character(center)) {
+    center <- which(rownames(dist) ==  center)[1]
+  }
+
   norm <- function(x){
+    if (!is.null(center)) {
+      x <- x - x[center]
+    }
     x <- abs(x)^power * sign(x)
-    (x-min(x))/(max(x)-min(x))
+    x[x>0] <- x[x>0]/max(x, na.rm = TRUE)
+    x[x<0] <- x[x<0]/(-min(x, na.rm = TRUE))
+    (x+1)/2
   }
   mds <- apply(mds, 2, norm)