+++ R/window.R
@@ -2,9 +2,8 @@
 # turn polygons from "maps" into "owin" window for spatstat
 # ====================
 
-mapsToOwin <- function(country, database = "worldHires") {
+mapsToOwin <- function(country, database = "world2Hires") {
 
-  requireNamespace("mapdata")
   raw <- maps::map(database = database
                    , regions = country
                    , plot = FALSE
@@ -22,7 +21,7 @@ mapsToOwin <- function(country, database = "worldHires") {
   }
   names(result) <- raw$names
 
-  return( spatstat::owin(poly = result) )
+  return( spatstat.geom::owin(poly = result) )
 }
 
 # ====================
@@ -31,18 +30,17 @@ mapsToOwin <- function(country, database = "worldHires") {
 
 gadmToOwin <- function(country, sub = NULL, level = 0) {
 
-  raw <- raster::getData("GADM", country = country, level = level)
+  raw <- geodata::gadm(country = country, level = level, path = tempdir())
 
   if (!is.null(sub)) {
-    name_objects <- which(grepl("^NAME_", names(raw@data)))
-    selection <- sapply(raw@data[name_objects], function(x) {
-      which(grepl(sub,x))
-    })
-    raw <- raw[unlist(selection),]
+    name <- paste("NAME", level, sep = "_")
+    d <- as.data.frame(raw)
+    col <- d[which(colnames(d) == name)]
+    selection <- which(col == sub)
+    raw <- raw[selection,]
   }
-
-  return( maptools::as.owin.SpatialPolygons(raw) )
-
+  raw <- spatstat.geom::as.owin(sf::as_Spatial(sf::st_as_sf(raw)))
+  return(raw)
 }
 
 # ====================
@@ -57,12 +55,12 @@ hullToOwin <- function(points, shift, alpha) {
                 , t(t(points) + c(-shift, shift))
   )
 
+  p <- sapply(p, jitter, factor = 1)
   hull <- alphahull::ahull(p, alpha =  alpha)
 
   # turn this into a "owin" window
-
   hull <- .ah2sp(hull)
-  hull <- maptools::as.owin.SpatialPolygons(hull)
+  hull <- polyCub::as.owin.SpatialPolygons(hull)
 
   return(hull)
 }
@@ -74,7 +72,7 @@ hullToOwin <- function(points, shift, alpha) {
 
 .ah2sp <- function(x
                   , increment = 360
-                  , rnd = 10
+                  , rnd = 5
                   , proj4string = sp::CRS(as.character(NA))
                   ) {