+++ NEWS.md
@@ -1,4 +1,18 @@
-# qlcData 0.1.0
+# qlcVisualize 0.2.1
+
+* bugfixes to get the package back on CRAN
+* trying to adapt to the ongoing changes in spatial methods/formats in R
+
+# qlcVisualize 0.2
+
+* selectively make pies in lmap only for those points that are multi-valued
+* adding options to normalize data in lmap
+* adding option 'ignore.others' in lmap
+* reorder plotting in lmap
+* allow for single row or column in limage
+* bigfixes in vmap and hullToOwin
+
+# qlcVisualize 0.1.0
 
 * initial alpha version on CRAN